% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tweets.R
\name{filter_tweets}
\alias{filter_tweets}
\title{Filter tweets}
\usage{
filter_tweets(data, keywords, include = TRUE)
}
\arguments{
\item{data}{Data frame containing tweets and hashtags. Works with any data frame, as long as there
is a "text" column of type character string and a "hashtags" column with comma separated character vectors.
Can be obtained either by using \code{\link{load_tweets}} on a json object returned by Twitter's API v1.1 or
by using \code{\link[jsonlite]{stream_in}} on any json file, as long as it has a "text" and "hashtags" field.
If you are unsure about the requirements you may load the sample piece of data contained in the package by
following the example in the the example section of this help page.}

\item{keywords}{Character string of keywords for black- or whitelisting provided via a comma separated character string.}

\item{include}{Logical. Indicate where to perform exclusive or inclusive filtering.
Inclusive filtering is akin to whitelisting keywords. Exclusive filtering is blacklisting certain keywords.}
}
\value{
Data frame of Tweets containing specified keywords
}
\description{
Filter tweets by keywords.
}
\details{
Use this function if you want your Tweets to contain certain keywords.
This can be used for iterative filtering to create more coherent topic models.
Keyword filtering is always case insensitive (lowercase).
}
\examples{
\dontrun{

library(Twitmo)


# load tweets (included in package)
mytweets <- load_tweets(system.file("extdata", "tweets_20191027-141233.json", package = "Twitmo"))

# Exclude Tweets that mention "football" and/or "mood"
keyword_dict <- "football,mood"
mytweets_reduced <- filter_tweets(mytweets, keywords = keyword_dict, include = FALSE)
}

}
