% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lda.R
\name{find_lda}
\alias{find_lda}
\title{Find best LDA model}
\usage{
find_lda(pooled_dfm, search_space)
}
\arguments{
\item{pooled_dfm}{object of class dfm (see \link[quanteda]{dfm}) containing (pooled) tweets}

\item{search_space}{Vector with number of topics to compare different models.}

\item{method}{The method to be used for fitting.
Currently method = "VEM" or method = "Gibbs" are supported.}
}
\value{
Plot with different metrics compared.
}
\description{
Find best LDA model
}
