% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stm.R
\name{find_stm}
\alias{find_stm}
\title{Find best STM model}
\usage{
find_stm(pooled_dfm, search_space)
}
\arguments{
\item{pooled_dfm}{object of class dfm (see \link[quanteda]{dfm}) containing (pooled) tweets}

\item{search_space}{Vector with number of topics to compare different models.}

\item{...}{Additional parameters passed to \link[stm]{searchK}}
}
\value{
Plot with different metrics compared.
}
\description{
Gridseaach for optimal K for your STM model. Wrapper function for \link[stm]{searchK}
}
\seealso{
\link[stm]{searchK}
}
