% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lda.R
\name{fit_lda}
\alias{fit_lda}
\title{Fit LDA Topic Model}
\usage{
fit_lda(pooled_dfm, n_topics)
}
\arguments{
\item{pooled_dfm}{Object of class dfm (see \link[quanteda]{dfm}) containing (pooled) tweets}

\item{n_topics}{Integer with number of topics}
}
\value{
Object of class \link[topicmodels:TopicModel-class]{LDA})
}
\description{
Estimate a LDA topic model using VEM or Gibbs Sampling.
}
