% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lda.R
\name{lda_distribution}
\alias{lda_distribution}
\title{View distribution of fitted LDA Models}
\usage{
lda_distribution(lda_model, param = "gamma", tidy = FALSE)
}
\arguments{
\item{lda_model}{Object of class \link[topicmodels:TopicModel-class]{LDA}).}

\item{param}{String. Specify either "beta" to return the term distribution.
over topics (term per document) or "gamma" for the document distribution over.
 topics (i.e. hashtag pool per topic probability).}

\item{tidy}{Boolean. Specify TRUE for return distribution in tidy format (tbl).}
}
\value{
Data frame or tbl of Term (beta) or document (gamma) distribution over topics.
}
\description{
View the distribution of your fitted LDA model.
}
