% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lda.R
\name{lda_hashtags}
\alias{lda_hashtags}
\title{View Documents (hashtags) heavily associated with topics}
\usage{
lda_hashtags(lda_model)
}
\arguments{
\item{lda_model}{Fitted LDA Model. Object of class \link[topicmodels:TopicModel-class]{LDA}).}
}
\value{
Data frame with most likely topic for each hashtag.
}
\description{
Convenience Function to extract the most likely topics for each hashtag.
}
