% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tweets.R
\name{load_tweets}
\alias{load_tweets}
\title{Converts Twitter stream data (JSON file) into parsed data frame}
\usage{
load_tweets(file_name, ...)
}
\arguments{
\item{file_name}{Character string. Name of JSON file with data collected by
\code{rtweet::stream_tweets()} or \code{get_tweets()}.}
}
\description{
This function has been deprecated in rtweet 1.0.0 but is
included here so ensure backwards compatibility data streamed with older versions of \code{rtweet}.
}
