% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hashtag}
\alias{plot_hashtag}
\title{Plot tweets with certain hashtag.}
\usage{
plot_hashtag(
  data,
  region = ".",
  alpha = 0.01,
  hashtag = "",
  ignore_case = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame of tweets parsed by \link[TweeLocViz]{load_tweets} or returned by \link[TweetLocViz]{pool_tweets}.}

\item{region}{Character vector specifying region. Returns a world \link[maps]{map} by default.
For higher resolutions specify a region.}

\item{alpha}{A double between 0 and 1 specifying the opacity of plotted points.
See \link[maps]{iso3166} for country codes.}

\item{hashtag}{Character vector of the hashtag you want to plot.}

\item{ignore_case}{Logical, if TRUE will ignore case of hashtag.}

\item{...}{Extra arguments passed to \link[graphics]{polygon} or \link[graphics]{lines}.}
}
\value{
Maps where each dot represents a tweet.
}
\description{
Plot the locations of certain hashtag on a static map with base plot.
}
\details{
This function can be used to generate high resolution spatial plots of tweets.
Works with data frames of tweets returned by \link[TweetLocViz]{pool_tweets} as well as data frames
read in by \link[TweetLocViz]{load_tweets} and then augmented by lat/lng coordinates with \link[rtweet]{lat_lng}.
For larger view resize the plot window then call \code{plot_tweets} again.
}
\examples{

\dontrun{

library(TweetLocViz)

# Plot hashtags on mainland USA
mytweets <- load_tweets("inst/extdata/tweets 20191027-141233.json")
plot_hashtag(mytweets, region = "USA(?!:Alaska|:Hawaii)", hashtag = "breakfast|chinup", ignore_case=TRUE, alpha=1)
# Add title
title("My Hashtags on a Map")
}

}
\seealso{
\link[maps]{map}, \link[maps]{iso3166}
}
