% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_tweets}
\alias{plot_tweets}
\title{Plot tweets on a map}
\usage{
plot_tweets(data, region = ".", alpha = 0.01, ...)
}
\arguments{
\item{data}{A data frame of tweets created by \link[TweetLocViz]{pool_tweets}.}

\item{region}{Character vector specifying region. Returns a world \link[maps]{map} by default.
For higher resolutions specify a region.}

\item{alpha}{A double between 0 and 1 specifying the opacity of plotted points.
See \link[maps]{iso3166} for country codes.}

\item{...}{Extra arguments passed to \link[graphics]{polygon} or \link[graphics]{lines}.}
}
\value{
Maps where each dot represents a tweet.
}
\description{
Visualize twitter data on a map.
}
\details{
This function can be used to generate high resolution spatial plots of tweets.
Works with data frames of tweets returned by \link[TweetLocViz]{pool_tweets} as well as data frames
read in by \link[TweetLocViz]{load_tweets} and then augmented by lat/lng coordinates with \link[rtweet]{lat_lng}.
}
\examples{

\dontrun{

library(TweetLocViz)

mytweets <- load_tweets("~/TweetsfromUK.json")
pool <- pool_tweets(mytweets)
plot_tweets(pool$data, region = "UK", alpha = 0.02, lwd = 0.1)
}

}
\seealso{
\link[maps]{map}, \link[maps]{iso3166}
}
