% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lda.R
\name{predict_lda}
\alias{predict_lda}
\title{Predict topics of tweets using fitted LDA model}
\usage{
predict_lda(
  data,
  lda_model,
  response = "max",
  remove_numbers = TRUE,
  remove_punct = TRUE,
  remove_symbols = TRUE,
  remove_url = TRUE,
  remove_separators = TRUE
)
}
\arguments{
\item{data}{Data frame of parsed tweets either by using `parse_stream()` or
\code{jsonlite::stream_in()} and \code{`rtweet::tweets_with_users(s)`}.}

\item{lda_model}{Fitted LDA Model. Object of class \link[topicmodels:TopicModel-class]{LDA}).}

\item{response}{Type of response. Either "prob" for probabilities or
"max" one topic (default).}
}
\value{
Data frame of topic predictions or predicted probabilities per topic (see response).
}
\description{
Predict topics of tweets using fitted LDA model.
}
