% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stm.R
\name{fit_stm}
\alias{fit_stm}
\title{Fit STM Topic Model}
\usage{
fit_stm(pooled_dfm, n_topics)
}
\arguments{
\item{pooled_dfm}{Object of class dfm (see \link[quanteda]{dfm}) containing (pooled) Tweets.}

\item{n_topics}{Integer with number of topics}

\item{...}{Additional arguments passed to \link[stm::stm]{stm}.}

\item{meta}{Optional data frame of external covariates i.e. meta data. By default uses favorite count,
retweet count, emojis and hashtags from Tweets.}
}
\value{
Object of class \link[stm:stm]{STM}
}
\description{
Estimate a STM topic model
}
