% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{get_tweets}
\alias{get_tweets}
\title{Sample tweets by streaming or searching}
\usage{
get_tweets(
  endpoint = "stream",
  bbox = c(-180, -90, 180, 90),
  sample_size = 100,
  location = NULL,
  ...
)
}
\arguments{
\item{endpoint}{One of two endpoints ('stream' or 'search'), which will call
upon `search_tweets()` or `stream_tweets()`.}

\item{bbox}{Use four latitude/longitude bounding box points to stream by geo
location. This must be provided via a vector of length 4, e.g., c(-125, 26, -65, 49).}

\item{sample_size}{Integer giving the total number of tweets to download when
using search endpoint.}

\item{location}{Locations are pre-defined bounding boxes.
Use \code{rtweet:::citycoords} to see a list.}

\item{...}{Other arguments passed to \code{stream_tweets()} or \code{search_tweets()}.}

\item{query}{Search query (optional). Use if you want to filter incoming tweets
by keyword. See \code{rtweet::stream_tweets() q} for more information.}
}
\value{
Either a json file in the specified directory, or (if `parse = TRUE`) additionally a data frame.
}
\description{
A wrapper function for convenient sampling of geo-tagged tweets
using either stream or search endpoints of Twitter API. Tweets can be searched
up to 9 days into the past.
}
