% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tweets.R
\name{load_tweets}
\alias{load_tweets}
\title{Converts Twitter stream data (JSON file) into parsed data frame}
\usage{
load_tweets(file_name)
}
\arguments{
\item{file_name}{Character string. Name of JSON file with data collected by
\link[rtweet]{stream_tweets} or \code{get_tweets()}.}
}
\description{
This function replaces \link[rtweet]{parse_stream} which has been
deprecated in rtweet > 0.7 but is included here to ensure backwards compatibility
for data streamed with older versions of \code{rtweet}.
Alternatively \code{jsonlite::streamin_in()} in conjunction with \code{rtweet::tweets_with_users()} and \code{rtweet::lat_lng()} can be used.
}
\seealso{
\link[rtweet]{parse_stream}.
}
