% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_tweets.R
\name{pool_tweets}
\alias{pool_tweets}
\title{Prepare for Tweets for topic modeling by pooling}
\usage{
pool_tweets(
  data,
  remove_numbers = TRUE,
  remove_punct = TRUE,
  remove_symbols = TRUE,
  remove_url = TRUE,
  remove_separators = TRUE
)
}
\arguments{
\item{data}{Data frame of parsed tweets.}
}
\value{
List with corpus object and dfm object of pooled tweets.
}
\description{
Pools tweets by hashtags using cosine similarity to create
longer pseudo-documents for better LDA estimation and creates n-gram tokens.
The method applies an implementation of the pooling algorithm from Mehrotra et al.2013.
}
\details{
This function pools parsed stream into hashtags.
}
