% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ldavis.R
\name{to_ldavis}
\alias{to_ldavis}
\title{Create interactive visualization with LDAvis}
\usage{
to_ldavis(fitted, corpus, doc_term).
}
\arguments{
\item{fitted}{Fitted LDA Model. Object of class \link[topicmodels:TopicModel-class]{LDA})}

\item{corpus}{Document corpus. Object of class \link[quanteda:corpus]{corpus})}

\item{doc_term}{document term matrix (dtm).}
}
\value{
Invisible Object (see \link[LDAvis]{serVis})).
}
\description{
Helper function. Converts topic model to LDAvis compatabile json string.
May requires \code{servr} to run properly.
}
